Added comments and enabled Emails on a visit saved.
