--- class/Report.php
@@ -1,24 +1,0 @@
-<?php
-namespace slc\reports;
-
-abstract class Report {
-
-
-	public function __construct();
-
-	public function execute();
-
-	public function getHtmlView();
-
-	public function getCsvView();
-
-
-
-
-	
-}
-
-?>
-}
-
-?>
