+++ class/ajax/POSTSendMail.php
@@ -1,30 +1,25 @@
 <?php 
+namespace slc\ajax;
 
 require_once PHPWS_SOURCE_DIR . 'lib/Swift/lib/swift_required.php';
 
 class POSTSendMail extends AJAX {
 
 	public function execute(){
-	
-		if ( !isset($_REQUEST['client_id']) ) {
-			throw new Exception('Client no longer in database.');
-			return;
-		}
 		
-		
-		$template = array("cName" => $_SESSION['cname']);
+		$template = array("cName" => $_REQUEST['name']);
 
-		$content = PHPWS_Template::process($template, 'slc', 'studentSurveyEmail.tpl');
+		$content = \PHPWS_Template::process($template, 'slc', 'studentSurveyEmail.tpl');
 
-		$banner = $_SESSION['actID'];
+		$banner = $_REQUEST['banner_id'];
 
 		$query = 'SELECT username FROM slc_student_data WHERE id =' . $banner;
 		
-		$db = new PHPWS_DB();
+		$db = new \PHPWS_DB();
 		$username = $db->select(null, $query);
 		$username = $username[0]['username'];
 		
-		$message = Swift_Message::newInstance();
+		$message = \Swift_Message::newInstance();
 
 		$message->setSubject('How was your experience with Student Legal Clinic?');
 		$message->setFrom('dos@appstate.edu');
@@ -32,14 +27,10 @@ class POSTSendMail extends AJAX {
 
 		$message->setBody($content);
 		
-		$transport = Swift_SmtpTransport::newInstance('localhost');
-		$mailer = Swift_Mailer::newInstance($transport);
+		$transport = \Swift_SmtpTransport::newInstance('localhost');
+		$mailer = \Swift_Mailer::newInstance($transport);
 
-		$mailer->send($message);
-		
-		//echo $username . '@appstate.edu';
-		//exit;
-				
+		$mailer->send($message);			
 	}
 }
