+++ javascript/viewClient/head.js
@@ -2,6 +2,7 @@
 <script src="https://fb.me/react-with-addons-0.13.1.js"></script>
 <script src="mod/intern/bower_components/react-bootstrap/react-bootstrap.min.js"></script>
 <script src="https://fb.me/JSXTransformer-0.13.1.js"></script>
+<script type = "text/jsx" src = "mod/slc/javascript/viewClient/ModalForm.jsx"></script>
 <script type = "text/jsx" src = "mod/slc/javascript/viewClient/ViewClientMain.jsx"></script>
 <script type = "text/javascript">
 var banner_id = {BANNER_ID};
